Move load() method from api.py to load.py
